Skip choosing view in the GUI if there is only 1 to choose.
